django_backend/questions/views.py
from  rest_framework import viewsets, permissions, filters
from django_filters.rest_framework import DjangoFilterBackend
from .models import Question
from .serializers import QuestionSerializer
from users.admin_views import IsAdminUser

class QuestionViewSet(viewsets.ModelViewSet):
    serializer_class = QuestionSerializer
    permission_classes = [permissions.IsAuthenticated, IsAdminUser]
    filter_backends = [DjangoFilterBackend, filters.SearchFilter]
    filterset_fields = ['specialization', 'year']
    search_fields = ['text']
    
    def get_queryset(self):
        return Question.objects.all().prefetch_related('options')
 