django_backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
python-decouple==3.8
psycopg2-binary==2.9.6
dj-database-url==2.0.0
whitenoise==6.4.0
gunicorn==20.1.0
 